+++ src/data/heroicons.js
@@ -7,23 +7,33 @@
 import arrowTopRightOnSquare from "heroicons/24/outline/arrow-top-right-on-square.svg?raw";
 import arrowTrendingUp from "heroicons/24/outline/arrow-trending-up.svg?raw";
 import bars3 from "heroicons/24/outline/bars-3.svg?raw";
+import bookOpen from "heroicons/24/outline/book-open.svg?raw";
 import briefcase from "heroicons/24/outline/briefcase.svg?raw";
 import cake from "heroicons/24/outline/cake.svg?raw";
 import faceSmile from "heroicons/24/outline/face-smile.svg?raw";
+import film from "heroicons/24/outline/film.svg?raw";
 import language from "heroicons/24/outline/language.svg?raw";
 import mapPin from "heroicons/24/outline/map-pin.svg?raw";
+import puzzlePiece from "heroicons/24/outline/puzzle-piece.svg?raw";
+import sparkles from "heroicons/24/outline/sparkles.svg?raw";
 import sun from "heroicons/24/outline/sun.svg?raw";
+import tv from "heroicons/24/outline/tv.svg?raw";
 import xMark from "heroicons/24/outline/x-mark.svg?raw";
 
 export const heroicons = {
   arrowTopRightOnSquare,
   arrowTrendingUp,
   bars3,
+  bookOpen,
   briefcase,
   cake,
   faceSmile,
+  film,
   language,
   mapPin,
+  puzzlePiece,
+  sparkles,
   sun,
+  tv,
   xMark,
 };